Updated usize to more concrete values.

+++ src/config.rs
@@ -1,10 +1,10 @@
-pub const MINE_RATE: usize = 1000; // milliseconds
-pub const MINE_RATE_DELTA: usize = 50;
+pub const MINE_RATE: u32 = 1000; // milliseconds
+pub const MINE_RATE_DELTA: u32 = 50;
 
-pub const GENESIS_TS: usize = 1;
+pub const GENESIS_TS: i64 = 1;
 pub const GENESIS_HASH: [u8; 4] = [1, 2, 3, 4];
 pub const GENESIS_LAST_HASH: [u8; 4] = [1, 2, 3, 4];
 pub const GENESIS_DATA: [&str; 1] = ["genesis_data"];
-pub const GENESIS_DIFFICULTY: usize = 5;
-pub const GENESIS_NONCE: usize = 0;
-pub const STARTING_BALANCE: usize = 1000;
+pub const GENESIS_DIFFICULTY: u32 = 5;
+pub const GENESIS_NONCE: u32 = 0;
+pub const STARTING_BALANCE: u32 = 1000;
